optimized GUI

+++ app/src/main/java/com/example/n008715/tipcalculator/MainActivity.java
@@ -58,7 +58,7 @@ public class MainActivity extends AppCompatActivity {
     private final OnSeekBarChangeListener seekBarListener = new OnSeekBarChangeListener() {
         @Override
         public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
-            percent = progress/100;
+            percent = (double)progress/100;
             calculate();
         }
